backend/pyproject.toml
[project]
name = "auracal-backend"
version = "0.1.0"
description = "AuraCal 灵光画布 - A Breathing, Persona-Driven Ambient Display"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.21.0",
    "openai>=1.60.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.19",
    "google-auth>=2.48.0",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]
